vplus-core/src/test/java/org/vplus/core/persistence/DBListTest.java
package org.vplus.core.persistence;

import static org.hamcrest.Matchers.equalTo;
import static org.junit.Assert.assertThat;

import java.util.List;

import org.dbunit.DatabaseUnitException;
import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import org.vplus.core.exeption.VPlusException;
import org.vplus.core.generics.Model;
import org.vplus.core.util.TestUtil;

public class DBListTest {

	DBList listDAO;
	private TestUtil testUtil;
	
	@Before
	public void setUp() throws Exception {
		testUtil = TestUtil.create();
		testUtil.from(MyEntity.class).init();
		listDAO = new DBList(testUtil.entityManager());
	}
	
	@After
	public void tearDown() throws DatabaseUnitException {
		testUtil.cleanAndDestroy();
	}
	
	@Test
	public void shouldReturn3ItemsFromDBUnit() throws VPlusException {
		List<Model> list = listDAO.of(MyEntity.class).find();
		assertThat(list.size(), equalTo(3));
	}
}